Avoid cyclig import issues when using ParseResults
