+++ pkg/services/dashboards/service/dashboard_service_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/grafana/grafana/pkg/storage/unified/search"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -1817,52 +1816,4 @@ func TestToUID(t *testing.T) {
 		result := toUID(rawIdentifier)
 		assert.Equal(t, "", result)
 	})
-}
-
-// regression test - parsing int32 values from search results was causing a panic
-func TestParseResults(t *testing.T) {
-	resSearchResp := &resource.ResourceSearchResponse{
-		Results: &resource.ResourceTable{
-			Columns: []*resource.ResourceTableColumnDefinition{
-				{
-					Name: "title",
-					Type: resource.ResourceTableColumnDefinition_STRING,
-				},
-				{
-					Name: "folder",
-					Type: resource.ResourceTableColumnDefinition_STRING,
-				},
-				{
-					Name: search.DASHBOARD_ERRORS_LAST_1_DAYS,
-					Type: resource.ResourceTableColumnDefinition_INT64,
-				},
-				{
-					Name: search.DASHBOARD_LINK_COUNT,
-					Type: resource.ResourceTableColumnDefinition_INT32,
-				},
-			},
-			Rows: []*resource.ResourceTableRow{
-				{
-					Key: &resource.ResourceKey{
-						Name:     "uid",
-						Resource: "dashboard",
-					},
-					Cells: [][]byte{
-						[]byte("Dashboard 1"),
-						[]byte("folder1"),
-						[]byte("100"),
-						[]byte("25"),
-					},
-				},
-			},
-		},
-		TotalHits: 1,
-	}
-
-	res, err := ParseResults(resSearchResp, 0)
-
-	require.NoError(t, err)
-	hitFields := res.Hits[0].Field.Object
-	require.Equal(t, int64(100), hitFields[search.DASHBOARD_ERRORS_LAST_1_DAYS])
-	require.Equal(t, int64(25), hitFields[search.DASHBOARD_LINK_COUNT])
 }
